requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
requests==2.31.0
anthropic>=0.40.0
groq>=0.9.0
python-dotenv==1.0.1
openpyxl==3.1.2
gunicorn==21.2.0
pandas==2.2.0
sqlalchemy==2.0.27
